nutrition_assistant_backend/pyproject.toml
[tool.poetry]
name = "nutrition-assistant-backend"
version = "0.1.0"
description = ""
authors = ["Matt Hendrick <75456865+matt-hendrick@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.107.2"
ollama = "^0.5.3"
pydantic = "^2.11.9"
tavily-python = "^0.7.12"
python-dotenv = "^1.1.1"
requests = "^2.32.5"
beautifulsoup4 = "^4.13.5"
fastapi = "^0.116.2"
uvicorn = "^0.35.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
